Fix circular import and _preprocess function
